bot/pyproject.toml
[project]
name = "lms-telegram-bot"
version = "0.1.0"
description = "Telegram bot for LMS analytics"
requires-python = ">=3.12"
dependencies = [
    "aiogram>=3.0.0",
    "python-dotenv>=1.0.0",
    "requests>=2.31.0",
    "openai>=1.0.0",
    "httpx>=0.27.0",
]

[dependency-groups]
dev = ["pytest>=8.0.0"]

[tool.poe.tasks]
bot = "python bot.py"
bot-test = "python bot.py --test"
